"put arreglado"

+++ SOLFM2K/Models/Rol.cs
@@ -5,13 +5,13 @@ namespace SOLFM2K.Models;
 
 public partial class Rol
 {
-    public byte RoEmpresa { get; set; }
+    public int RoEmpresa { get; set; }
 
-    public short RoCodigo { get; set; }
+    public int RoCodigo { get; set; }
 
     public string? RoNombre { get; set; }
 
     public string? RoEstado { get; set; }
 
-    public short? RoAplicacion { get; set; }
+    public int? RoAplicacion { get; set; }
 }
